feat: 新增了聊天室加载历史消息和实时显示历史对话的功能，解决了用户退出登录后，再次登录发送信息，自己聊天窗口无法显示自己发送信息的并发bug问题，存在的已知bug，redis 6.0.16执行开区间的stream命令失败

+++ server/application/chat-room/api/api_common.cc
@@ -81,7 +81,7 @@ string api_error_id_to_string(api_error_id input) {
 /**
  * query redis
  */
-int ApiGetUsernameAndUseridByCookie(string cookie, string& username, int32_t& userid, string& email) {
+int ApiGetUsernameAndUseridByCookie(string cookie, string& username, int64_t& userid, string& email) {
     int ret = 0;
     CacheManager* cache_manager = CacheManager::getInstance();
     CacheConn* cache_conn = cache_manager->GetCacheConn("token");
@@ -105,7 +105,7 @@ int ApiGetUsernameAndUseridByCookie(string cookie, string& username, int32_t& us
 /**
  * query mysql
  */
-int GetUserNameAndUseridByEmail(string& email, string& username, int32_t& userid) {
+int GetUserNameAndUseridByEmail(string& email, string& username, int64_t& userid) {
     int ret = 0;
     CDBManager* db_manager = CDBManager::getInstance();
     CDBConn* db_conn = db_manager->GetDBConn("chatroom_slave");
